+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+Added
+  * AZ-1194: Add init_containers block
+  * AZ-1201: Add dns_config block
+
 # v7.4.0 - 2023-09-29
 
 Breaking
@@ -7,7 +13,6 @@ Added
   * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add liveness and readiness probes
   * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add DNS name label reuse policy
   * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add identity block
-  * AZ-1194: Add init_containers block
 
 # v7.3.0 - 2022-12-09